refactor: simplify ClassGenerator orchestration
